requirements.txt
# PDF Processing
llama-parse==0.6.62

# LLM Integration
google-generativeai==0.8.5

# Data Processing
pandas==2.3.3
numpy==2.3.4

# Environment Variables
python-dotenv==1.2.1

# Embeddings and Vector Search (for bonus features)
sentence-transformers==2.2.2
faiss-cpu==1.7.4
scikit-learn==1.3.212.0

# Utilities
tqdm==4.66.1
pathlib==1.0.1
json5==0.9.14

# Web Interface
flask==3.0.0
werkzeug==3.0.1

# LlamaIndex Core (compatible with llama-parse 0.6.62)
llama-index-core>=0.12.0,<0.15.0